+++ sykepenger-model/src/main/kotlin/no/nav/helse/utbetalingstidslinje/Alder.kt
@@ -3,8 +3,7 @@ package no.nav.helse.utbetalingstidslinje
 import no.nav.helse.Grunnbeløp
 import java.time.DayOfWeek
 import java.time.LocalDate
-import java.time.temporal.ChronoUnit
-import java.time.temporal.ChronoUnit.*
+import java.time.temporal.ChronoUnit.YEARS
 
 internal class Alder(fødselsnummer: String) {
     private val individnummer = fødselsnummer.substring(6, 9).toInt()
@@ -38,6 +37,6 @@ internal class Alder(fødselsnummer: String) {
     internal fun alderPåDato(dato: LocalDate) = YEARS.between(fødselsdag, dato).toInt();
 
     internal fun minimumInntekt(dato: LocalDate): Double {
-        return (if (dato <= redusertYtelseAlder) Grunnbeløp.halvG else Grunnbeløp.`2G`).dagsats(dato)
+        return (if (dato <= redusertYtelseAlder) Grunnbeløp.halvG else Grunnbeløp.`2G`).dagsats(dato).toDouble()
     }
 }
